+++ jest.config.json
@@ -1,14 +1,7 @@
 {
   "preset": "ts-jest",
   "testRegex": "(/tests/.*(test|spec))\\.tsx?$",
-  "transformIgnorePatterns": ["/node_modules/"],
-	"coverageThreshold": {
-    "global": {
-      "branches": 10,
-      "functions": 10,
-      "lines": 10,
-      "statements": 10
-    }
-  },
-  "collectCoverageFrom": ["./src/**/*.{js,jsx,ts,tsx}", "!**/tests/**"]
+	"moduleNameMapper": {
+		"/src/(.*)": "<rootDir>/src/$1"
+	}
 }
