+++ package.json
@@ -1,49 +1,49 @@
 {
-  "name": "make-pot",
-  "version": "0.0.1",
-  "description": "A Node.js script for generating a POT file from source code",
-  "author": "Erik Golinelli <https://github.com/erikyo>",
-  "scripts": {
-    "build": "tsc --build",
-    "watch": "tsc --watch",
-    "clean": "tsc --build --clean",
-    "lint": "eslint ./src ./tests --fix",
-    "test": "jest --coverage",
-    "run": "node ./lib/index.js tests/fixtures/ --exclude null --include \"**\""
-  },
-  "bin": {
-    "make-pot": "lib/index.js"
-  },
-  "dependencies": {
-    "cli-progress": "^3.12.0",
-    "gettext-parser": "^7.0.1",
-    "glob": "^10.3.10",
-    "yargs": "^17.7.2"
-  },
-  "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "@types/cli-progress": "^3.11.5",
-    "@types/gettext-parser": "^4.0.4",
-    "@types/jest": "^29.5.11",
-    "@types/node": "^20.10.4",
-    "@types/yargs": "^17.0.32",
-    "@typescript-eslint/eslint-plugin": "^6.18.1",
-    "@typescript-eslint/parser": "^6.14.0",
-    "eslint": "^8.56.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-config-standard-with-typescript": "^43.0.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-n": "^16.6.1",
-    "eslint-plugin-promise": "^6.1.1",
-    "husky": "^8.0.3",
-    "jest": "^29.7.0",
-    "lint-staged": "^15.2.0",
-    "prettier": "3.2.2",
-    "tree-sitter": "^0.20.6",
-    "tree-sitter-javascript": "^0.20.1",
-    "tree-sitter-php": "^0.20.0",
-    "tree-sitter-typescript": "^0.20.3",
-    "ts-jest": "^29.1.1",
-    "typescript": "^5.3.3"
-  }
+	"name": "make-pot",
+	"version": "0.0.1",
+	"description": "A Node.js script for generating a POT file from source code",
+	"author": "Erik Golinelli <https://github.com/erikyo>",
+	"scripts": {
+		"build": "tsc --build",
+		"watch": "tsc --watch",
+		"clean": "tsc --build --clean",
+		"lint": "eslint ./src ./tests --fix",
+		"test": "jest --coverage",
+		"run": "node ./lib/index.js tests/fixtures/ --exclude null --include \"**\""
+	},
+	"bin": {
+		"make-pot": "lib/index.js"
+	},
+	"dependencies": {
+		"cli-progress": "^3.12.0",
+		"gettext-parser": "^7.0.1",
+		"glob": "^10.3.10",
+		"yargs": "^17.7.2"
+	},
+	"devDependencies": {
+		"@jest/globals": "^29.7.0",
+		"@types/cli-progress": "^3.11.5",
+		"@types/gettext-parser": "^4.0.4",
+		"@types/jest": "^29.5.11",
+		"@types/node": "^20.10.4",
+		"@types/yargs": "^17.0.32",
+		"@typescript-eslint/eslint-plugin": "^6.18.1",
+		"@typescript-eslint/parser": "^6.14.0",
+		"eslint": "^8.56.0",
+		"eslint-config-standard": "^17.1.0",
+		"eslint-config-standard-with-typescript": "^43.0.0",
+		"eslint-plugin-import": "^2.29.1",
+		"eslint-plugin-n": "^16.6.1",
+		"eslint-plugin-promise": "^6.1.1",
+		"husky": "^8.0.3",
+		"jest": "^29.7.0",
+		"lint-staged": "^15.2.0",
+		"prettier": "3.2.2",
+		"tree-sitter": "^0.20.6",
+		"tree-sitter-javascript": "^0.20.1",
+		"tree-sitter-php": "^0.20.0",
+		"tree-sitter-typescript": "^0.20.3",
+		"ts-jest": "^29.1.1",
+		"typescript": "^5.3.3"
+	}
 }
