--- tests/fixtures/file1.txt
@@ -1,1 +1,0 @@
-file1.txt
