--- tests/fixtures/fixtures.php
@@ -1,22 +1,0 @@
-/**
- * Fixtures
- *
- * @package           fixtures
- * @author            Your Name
- * @copyright         2019 Your Name or Company Name
- * @license           GPL-2.0-or-later
- *
- * @wordpress-plugin
- * Plugin Name:       Plugin Name
- * Plugin URI:        https://example.com/plugin-name
- * Description:       Description of the plugin.
- * Version:           1.0.0
- * Requires at least: 5.2
- * Requires PHP:      7.2
- * Author:            Your Name
- * Author URI:        https://example.com
- * Text Domain:       plugin-slug
- * License:           GPL v2 or later
- * License URI:       http://www.gnu.org/licenses/gpl-2.0.txt
- * Update URI:        https://example.com/my-plugin/
- */
